requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.14.0
pillow==10.3.0
platformdirs==4.2.2
psycopg2==2.9.9
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.2